refactor : 디렉토리 리펙토링, (TODO: ts 문법 적용해야함)
