--- .idea/shelf/Uncommitted_changes_before_Update_at_2022_02_13_17_03__Changes_.xml
@@ -1,4 +1,0 @@
-<changelist name="Uncommitted_changes_before_Update_at_2022_02_13_17_03_[Changes]" date="1644739411370" recycled="true" deleted="true">
-  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Update_at_2022_02_13_17_03_[Changes]/shelved.patch" />
-  <option name="DESCRIPTION" value="Uncommitted changes before Update at 2022/02/13 17:03 [Changes]" />
-</changelist>
