+++ pages/_app.tsx
@@ -1,7 +1,7 @@
 import type { AppProps } from 'next/app'
 import { Provider } from 'react-redux'
 // import { store } from '../src/store'
-import {store} from '../src/redux/store'
+import {store} from '../src/features/store'
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
