+++ src/features/store.ts
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
-import TodoReducer from './todoRoom/Todos'
-import DoingReducer from './todoRoom/Doings'
-import DoneReducer from './todoRoom/Dones'
+import TodoReducer from './todoRoom/redux_todoRoom/Todos'
+import DoingReducer from './todoRoom/redux_todoRoom/Doings'
+import DoneReducer from './todoRoom/redux_todoRoom/Dones'
 
 export const store = configureStore({
   reducer: {
